requirements.txt
clingo==5.5.0.post3
pygame==2.0.1
py-sudoku==1.0.1
